FR-19895: Added app comments
