+++ app/src/main/java/com/frontegg/demo/ui/tenants/TenantsViewModel.kt
@@ -10,7 +10,7 @@ import com.frontegg.android.models.Tenant
 
 class TenantsViewModel : ViewModel() {
 
-
+    // LiveData to hold the list of tenants associated with the authenticated user
     private val _tenants = MutableLiveData<List<Tenant>>().apply {
         FronteggAuth.instance.user.subscribe {
             Handler(Looper.getMainLooper()).post {
@@ -19,6 +19,7 @@ class TenantsViewModel : ViewModel() {
         }
     }
 
+    // LiveData to hold the currently active tenant for the authenticated user
     private val _activeTenant = MutableLiveData<Tenant?>().apply {
         FronteggAuth.instance.user.subscribe {
             Handler(Looper.getMainLooper()).post {
